// client/webapi/models/login-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Barber App API
 * An API for Barber App 
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface LoginRequest
 */
export interface LoginRequest {
    /**
     * 
     * @type {string}
     * @memberof LoginRequest
     */
    email?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LoginRequest
     */
    password?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LoginRequest
     */
    twoFactorCode?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LoginRequest
     */
    twoFactorRecoveryCode?: string | null;
}