// client/webapi/models/access-token-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Barber App API
 * An API for Barber App 
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AccessTokenResponse
 */
export interface AccessTokenResponse {
    /**
     * 
     * @type {string}
     * @memberof AccessTokenResponse
     */
    tokenType?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AccessTokenResponse
     */
    accessToken?: string | null;
    /**
     * 
     * @type {number}
     * @memberof AccessTokenResponse
     */
    expiresIn?: number;
    /**
     * 
     * @type {string}
     * @memberof AccessTokenResponse
     */
    refreshToken?: string | null;
}